set number of reducers to reflect maximum parrallelism possible

+++ com.cloudera.companies.core/com.cloudera.companies.core.common/src/main/java/com/cloudera/companies/core/common/mapreduce/CompaniesFileKeyGroupPartitioner.java
@@ -7,7 +7,7 @@ public class CompaniesFileKeyGroupPartitioner extends Partitioner<CompaniesFileK
 
 	@Override
 	public int getPartition(CompaniesFileKey key, Text value, int numPartitions) {
-		return (int) ((((long) key.getGroup().hashCode()) & 0x00000000ffffffffL) % numPartitions);
+		return (int) (((key.getGroup().hashCode()) & 0x00000000ffffffffL) % numPartitions);
 	}
 
 }
